Write a NilCache test, better error documentation
NilCache really needn’t be a type; there is only one, and it may as well
be the value `(*nilCache)(nil)`. Its type is `*nilCache` (not `Cache`),
so reassigning the value should at least not affect the cache behavior.

After some thought, I decided that ErrNotStored is fine to return from
NilCache’s Add method; the reasoning is explained in a comment.

This also led to documenting the reexported gomemcache errors, writing
down the assumptions I’d been implicitly making.

This also adds an actual test case for NilCache, demonstrating that it
calls the raw function every time and produces no non-fatal errors under
normal circumstances.